Move validation function to a separate file
